Implemented the storage auth data using the config
